Refactored class
